Make project versioning clearer (#146)
This is now managed purely by the CD when a GitHub release is created, so the specific version of the package no longer needs to be checked in. When the package is releases to PyPi, this version will be set correctly.

Also adds a badge to the readme so that we keep visibility as to the version.