Fix FactoredMatrix bug (#367)
* Fix FactoredMatrix bug

The function `_convert_to_slice` was added recently, which causes indexing with elements to work (previously this didn't work). But when this change was made, it caused indexing with sequences to work. The solution is to make `_convert_to_slice` only alter integer elements of the index.

To further clarify, `_convert_to_slice` will take an integer like `1` and convert it into the slice `slice(1, 2)`, this way the indexed shapes have the correct number of dimensions. But this isn't how we should treat things like `t.tensor([0, 1, 2])` when they're used as an index, because they're already in the correct form for indexing (i.e. indexing with them won't delete a dimension).

* Added tests for sequences

I've added tests which would have broken on the previous edit to FactoredMatrices, but which now pass.

* Fix formatting and torch

* Remove line break for formatting