Fix attention result projection (#666)
* Updated README to have Bryce as the maintainer

* Fix attention result projection

Current result projection for attention is incorrect. Type annotations would suggest that `result` isn't being summed over `head_index`, but in fact it is. I've edited the function so that it's no longer being summed over `head_index`.

Note, this bug caused the ARENA material to fail for the first transformers chapter, I've tested it and it now works.

* fix formatting with black

---------

Co-authored-by: Bryce Meyer <bryce13950@gmail.com>
Co-authored-by: Neel Nanda <neelnanda27@gmail.com>